// packages/browser/common/src/types/app.ts
export enum EApplicationName {
    Settings = 'settings',
    Dashboard = 'dashboard',
    BacktestingManager = 'backtestingManager',
    TradingServersManager = 'tradingServersManager',
    TradingStats = 'tradingStats',
    PortfolioTracker = 'portfolioTracker',
    HerodotusDashboard = 'herodotusDashboard',
    HerodotusTerminal = 'herodotusTerminal',
    HerodotusTrades = 'herodotusTrades',
    BalanceMonitor = 'balanceMonitor',
    CharterJsonViewer = 'charterJsonViewer',
    WSQueryTerminal = 'wsQueryTerminal',
}